python-backend-for-ai/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
chromadb
pydantic
python-dotenv
